+++ import-service/handlers/import-products-file/import-products-file.ts
@@ -1,8 +1,8 @@
 import {APIGatewayProxyHandler} from 'aws-lambda';
 import 'source-map-support/register';
 import * as AWS from 'aws-sdk';
-import {getRes200, getRes400, getRes500, logLambdaArgs, messages} from '../../shared/utils'
-import {S3_BUCKET_NAME} from '../utils';
+import {getRes200, getRes400, getRes500, logLambdaArgs, messages} from '../../../shared/utils'
+import {S3_BUCKET_NAME} from '../../utils';
 
 export const getS3Params = (csvFileName) => ({
   Bucket: S3_BUCKET_NAME,
