+++ import-service/handlers/index.ts
@@ -1,0 +1,4 @@
+import { importProductsFile } from './import-products-file/import-products-file';
+import { importFileParser } from './import-file-parser/import-file-parser';
+
+export { importProductsFile, importFileParser };
