+++ products-service/handlers/index.ts
@@ -1,0 +1,5 @@
+import { createProduct } from './create-product';
+import { getAllProducts } from './get-all-products';
+import { getProductById } from './get-product-by-id';
+
+export { createProduct, getAllProducts, getProductById };
